# Performance optimization