Hi again again @m1l,

Thanks for merging my previous PRs so diligently! I have yet another one for you (I _do_ try to split them in smaller batches!). Sorry if it is becoming harassment at this point, do feel tree to take your time, I am in no hurry ;-) .

In any case, the previous PRs have laid the groundwork by adding testing and starting to make some parts of the code look more uniform. Maybe more importantly, it has also helped me understand better how the extension work so that I can iterate on your work without breaking too many things.

This time, I focus on performance. With your [remark on #38](https://github.com/m1l/Everything-Metric-Firefox/issues/38#issuecomment-1749693670), I realized I had not performed any benchmarking, assuming that the code was fast enough for all practical purposes. Additionally, I think my previous refactors did help make the code more robust and fix a few minor bugs, but they did make things slower. This is how I proceeded to measure performance:

1. Open the test page `index.html` in a clean Firefox instance
2. Open the Performance tab in the dev tools
3. Recording the performance for a hundred page-reloads (I did that reliably with `xdotool click 1; sleep 1; xdotool key --repeat 100 --delay 500 F5 click 1`)
4. In the flame graph, focus on the node `(root scope) moz-extension://36455235-70f9-4c2e-b657-3c25cbcf3e1e/contentscript.js` to only look at the code from the extension

This is the total runtime (divided by 100) for different versions of the extensions:

- v3.5: 62 ms
- current `master` (84f1a6faaef8f2369f8f650f48d1c73c5434dab0): 125 ms
- this PR: 57 ms
- [`remove-useComma-useSpaces`](https://github.com/m1l/Everything-Metric-Firefox/commit/5497443c1e09a413055391c79c9102231b4e8355#diff-4d4d0cea7e443590b59dddca4fc34faa6f37fbfd94a251d8d0047b8aa8ca19eeR449): 36 ms

So we can see that my previous refactors did make things much slower. But this PR brings things back to manageable territory. Additionally, dropping the `useComma` and `useSpaces` feature, and just using `value.toLocaleString()` without any arguments represents a significant gain in performance. I actually think that would suit most users better than manual options, but it could be considered a breaking change, so I have not included it in this PR. I would like your opinion on this one!

To perform this improvement, I have looked at the call tree with “inverted call stack”, which shows in what functions the code spends most of his time. In the current `master`, it looks like.

![image](https://github.com/m1l/Everything-Metric-Firefox/assets/8493765/7ef26a7a-d5dd-40f9-a1c4-85fab5452ab0)

With this PR, it looks like:

![image](https://github.com/m1l/Everything-Metric-Firefox/assets/8493765/cc9a3901-7f82-401c-8403-0752ef2d02ef)

Before doing this profiling, I actually expected that most time would be spent in `insertAt` since it replaces the string for the whole text at each replacement, but it is actually a rounding error, at just 0.1 %! But, as you can see, most of the time is spent in the `RegExp.exec` function, which does make sense, since this is the function that does the heavy lifting for the conversion logic.

The first 20 commits mostly clean up the regexes for the unit conversions to make them more alike

I spent some time trying to optimize the regexes themselves (in particular `parseNumberPattern` since it is so complex, is used everywhere, and its interaction with following patterns could theoretically cause a quadratic blowup with spaces), but I did not find any significant difference, except for one change, which is dropping the Unicode support and `\p{Nd}` class. Since I introduced it myself in https://github.com/m1l/Everything-Metric-Firefox/pull/37/commits/0441a5ea3cebdfb5abc13f1a0cf6e9d186c57a27 to parse full-width digits, and since it would be improbably to see full-width digits (mostly used in Eastern languages) with US customary units, I think there is little point in keeping them. So it is removed by cc9668d48f1347ca9c2cceb66c86dee78098457a.

Instead, I focused on trying to reduce the number of times `RegExp.exec` is called. I managed to do it to an extend with a simple trick: instead of having `replaceOtherUnits` iterate over each `Conversion` and search the text for each instance of the conversion, I wrote a single regex that would find all such conversions. Then, for each instance, I just use `RegExp.test` to find out which conversion should be performed. This is performed in 85e044bb659744f3c55fdc4114981ae3cdc845ef, and brings the most performance gain. It also moves us closer to merging `Conversion.regex` and `Conversion.regexUnit`, which would be nice!

Additionally, I made sure that regexes would only be built once, with 0e62cc0bff4522b577aaebb52f1af1544b27991f and f0407c8a70377407ff6121c4f1462ec162347aa7. It does not look as big in the screenshot from `master`, but `RegExp constructor` and `Array.prototype.join` start taking a non-trivial amount of relative time after the previous optimization. I have also optimized the `maketrans` & `String.prototype.translate` in b536ccf734f8f294de5ed0335d6bf74f07593a13; it's a minor gain, but pretty consistent, whether full-width digits are supported or not.

In any case, the most critical change is in 85e044bb659744f3c55fdc4114981ae3cdc845ef, and it is also the more likely to bring regressions, so we should scrutinize the changes in oracles in this commit. Some changes are actually even improvement (it fixes #26!), but it actually breaks `born in 1948 in`, and the change from `1 yd (91 cm)˜ 2 in (0,97 m)˜` to `1 yd 2 in (0.97 m)˜` could lead to confusion. Since `1 yd` is just ignored.

I am opening this PR, so you can start looking at the changes, and so we can discuss whether dropping support for `useComma` and `useSpaces` would be worth it, but I should fix the small issues I discussed in the previous paragraph before this is merged.

And again, your code is pretty impressive, with all the edge cases it handles!