@qsantos 2023-10-09
Okay, the oracle change regarding `born in 1948 in` is actually just restoring the behavior from before 0441a5ea3cebdfb5abc13f1a0cf6e9d186c57a27, so it's actually okay.

The one with `1 yd 2 in (0.97 m)˜` comes from the fact that `unitSuffixInFt`, which is meant to avoid converting feet followed by inches, is now applied to all units instead of just feet. Making this correct in all cases is non-trivial. If that's okay with you, I would like to keep it as-is for this PR. We can open an issue to remember to handle this case if you think that's worth it.