@m1l 2023-10-15
I am not sure about replacing the useComma and useSpaces feature with value.toLocaleString(). It sounds reasonable but even though I come from culture using comma for decimal separator, and had to adjust to US and software dev world using . as decimal separator, I don't know how regular people or expats look at these (or if they even care). My guess is that value.toLocaleString() is probably fine for 99% of people. 

I see you have fixed some of those unnecessary conversions, great!

One thing that I am not sure if was obvious - the reason why some units were completely replaced and others had conversion inserted after the imperial unit is the improvisation so that a multi-unit measurement wouldn't be parsed once, and then again for second time for individual ft+in conversion.

I like the performance benchmarks you did. Since you already have it set this up, can you please check if it also makes a difference in other long texts regardless if there is anything to convert, like a very long article or an [ebook ](https://www.gutenberg.org/cache/epub/1513/pg1513-images.html) that someone will open sooner or later?