@qsantos 2023-10-15
Thanks for reviewing and merging this!

> I am not sure about replacing the useComma and useSpaces feature with value.toLocaleString(). It sounds reasonable but even though I come from culture using comma for decimal separator, and had to adjust to US and software dev world using . as decimal separator, I don't know how regular people or expats look at these (or if they even care). My guess is that value.toLocaleString() is probably fine for 99% of people.

Note that `value.toLocaleString()` _does_ take into account the user's locale. On my Firefox profile set in French:
![image](https://github.com/m1l/Everything-Metric-Firefox/assets/8493765/b2655cb4-2add-492a-9919-6714b8e55cae)

So the change would not be about not localizing the numbers, but instead just using the global browser's settings, instead of changing them locally. From what I understand, the performance gain comes from the fact that the current user's locale is immediately accessible, while passing e.g. `'de-DE'` as an argument implies a look-up at each invocation.

If you think it would still be worthwhile to keep the behavior of `useComma` and `useSpaces`, we could do something like this:
- if `useComma` or `useSpaces` is not set to its default value (`true`), then keep the behavior with manual formatting
- if they are set to their default values, just call `.toLocaleString()`

> One thing that I am not sure if was obvious - the reason why some units were completely replaced and others had conversion inserted after the imperial unit is the improvisation so that a multi-unit measurement wouldn't be parsed once, and then again for second time for individual ft+in conversion.

I understand. I haven't thought of a simple solution, but it would be nice to do all conversions in one pass. That would make the regex more complicated, but we might still win from not recognizing numbers several times, for instance.

> I like the performance benchmarks you did. Since you already have it set this up, can you please check if it also makes a difference in other long texts regardless if there is anything to convert, like a very long article or an [ebook ](https://www.gutenberg.org/cache/epub/1513/pg1513-images.html)that someone will open sooner or later?

There does not seem to be many US customary units, but the extension does take a significant amount of time: 363 ms before this PR, and still 130 ms after this PR. However, it looks like it's now mostly in the DOM walking logic:

![image](https://github.com/m1l/Everything-Metric-Firefox/assets/8493765/3704b253-a9fe-4666-af22-bb85df4e96ec)

There might still there be something to do about that. I will take a look.

For a document with many required conversions, we can use an English-translation of Jules Verne. I have tested both the [Web version](https://www.gutenberg.org/cache/epub/83/pg83-images.html) and the [text version](https://www.gutenberg.org/cache/epub/83/pg83.txt) of the combined _From the Earth to the Moon_ & _Round the Moon_.

In the Web version, it takes 165 ms, and we see a similar pattern as in Romeo and Juliet, with most of the time spend in the DOM, although there is a significant amount of time (> 20 %) in text processing:

![image](https://github.com/m1l/Everything-Metric-Firefox/assets/8493765/064c1ee1-7dd5-4b4a-8d4f-cb33fdb02596)

In the text version, you might expect the runtime to be shorter. However, we are actually at 264 ms, even without spending any significant time in the DOM. It is probably due to handling the text as a whole, instead of piece by piece. I suspect that this is due to the fact that we make a copy of the full text for each conversion, but I need to confirm that.

![image](https://github.com/m1l/Everything-Metric-Firefox/assets/8493765/0f3bf534-1ac5-49a5-bb4b-110b9f45e2e3)