@m1l 2023-10-21
`toLocaleString` is probably doing everything we need. I was not aware of `toLocalString` when I created this or I probably wouldn't have used those options.

I lean toward simpler regex even if it costs some performance for easier bug-fixing in the future. In the past I thought of looking for numbers block first and grabbing few words before and after, then try to match if those were units. That may be faster as the first pass, but then it becomes solving several problems at the same time. This was still fast enough (and with your findings sounds it is even faster so it may not be necessary or beneficial in real life)